+++ Domain/Carts/CartItem.cs
@@ -6,7 +6,7 @@ namespace PersonalShop.Domain.Card;
 
 public class CartItem
 {
-    public CartItem(long productId,int quanity)
+    public CartItem(int productId,int quanity)
     {
         ProductId = productId;
         Quanity = quanity;
@@ -16,7 +16,7 @@ public class CartItem
     public int Id { get; set; }
     public Guid CartId { get; set; }
     public Product Product { get; set; } = null!;
-    public long ProductId { get; private set; }
+    public int ProductId { get; private set; }
     public int Quanity { get; private set; }
 
     public void IncreaseQuantity(int count)
